+++ CHANGES.rst
@@ -4,6 +4,8 @@ Revision 0.1.5, released XX-07-2018
 
 - The `noescape` option implemented
 - Allow empty value syntax in the option-value pair e.g. `option: `
+- Fixed a bug causing malformed AST when a comment resides
+  in-between duplicate keys
 
 Revision 0.1.4, released 27-05-2018
 -----------------------------------